+++ src/rex.forms/src/rex/forms/ctl.py
@@ -10,7 +10,8 @@ from rex.core import Error, AnyVal
 from rex.ctl import Task, RexTask, argument, option
 from rex.instrument import InstrumentVersion
 from rex.instrument.ctl import \
-    open_and_validate as open_and_validate_instrument
+    open_and_validate as open_and_validate_instrument, \
+    ImplementationContextReceiver
 from rex.instrument.util import get_implementation
 
 from .errors import ValidationError
@@ -265,7 +266,7 @@ class FormsRetrieveTask(RexTask, FormOutputter):
             self.do_output(form.configuration)
 
 
-class FormsStoreTask(RexTask):
+class FormsStoreTask(RexTask, ImplementationContextReceiver):
     """
     stores a Form in the data store
 
@@ -351,6 +352,10 @@ class FormsStoreTask(RexTask):
                     channel,
                     instrument_version,
                     configuration,
+                    implementation_context=self.get_context(
+                        form_impl,
+                        form_impl.CONTEXT_ACTION_CREATE,
+                    ),
                 )
                 print 'Created new Form'